Extract HTMX action logic for fledgling test util
